check if files need removeal

+++ custom/bin/backup_home.sh
@@ -79,7 +79,9 @@ if [ "$1" = "-l" ];then
 	backup_file_prefix="$(cd $(dirname "$BACKUP_DIR");pwd)/$(basename "$BACKUP_DIR")"
 	backup_file=${backup_file_prefix}_$(date +%F.%H%M%S).tgz
 	tar --use-compress-program=pigz -c -f $backup_file -C $BACKUP_BASE $BACKUP_NAME > /dev/null
-	ls -t ${backup_file_prefix}_*.tgz |tail -n +2 | xargs rm --
+	if ls ${backup_file_prefix}_*.tgz 1>/dev/null 2>&1;then
+		ls -t ${backup_file_prefix}_*.tgz |tail -n +2 | xargs rm --
+	fi
 	echo "done. [$backup_file]" 
 
 	ENCRYPT_DIR="$BACKUP_BASE/encrpyted_backups"
